Boss spawn but it's all multiplied by 4
idk

+++ battle/entity/AbEntity.java
@@ -18,7 +18,7 @@ public abstract class AbEntity extends BattleObj {
 		health = maxH = h;
 	}
 
-	public void added(int d, int p) {
+	public void added(int d, double p) {
 		pos = p;
 		dire = d;
 	}
